Add reinstall flag to language pack installs
The check-language-support and apt are not the same opinion on whether
the package is missing or not. check-language-support identifies
packages as missing and apt says they are already installed. In the
final installed system the packages, that have been identified as
"installed" by apt, are missing. By reinstalling I hope to force apt to
actually do something and not skip the packages.